Create RHOSRFE-61_new_cli_flags
This change would create several new CLI flags: current-config,
rollback, no-abort, and add functionality to detect ip duplicates
when using --noop.
--current-config
Be able to remove interfaces without using the
overly disruptive --cleanup option, and pass os-net the existing
configuration file as well as the new one so that a delta may be
calculated and applied accordingly.
--rollback
In the event of failure to apply the network, revert back to the
previous configuration.
--no-abort
Continue to try and apply the rest of the configuration, even if one
interface fails, and exit with a special return code in this case.
--noop
Change behavior so that when using --noop os-net-config would
attempt to detect if any of the IP addresses are already in use on
hosts other than the current host. This would allow a trial run to
predict whether any IP conflicts would occur during a production
run.